+++ src/drivers/imu/bosch/bmi088/BMI088_Accelerometer.cpp
@@ -346,7 +346,7 @@ int BMI088_Accelerometer::DataReadyInterruptCallback(int irq, void *context, voi
 
 void BMI088_Accelerometer::DataReady()
 {
-	uint8_t expected = 0;
+	uint32_t expected = 0;
 
 	if (_drdy_fifo_read_samples.compare_exchange(&expected, _fifo_accel_samples)) {
 		ScheduleNow();
